# requirements/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CarsAI API
  version: 0.1.0
  description: |
    REST API контракт для CarsAI.

    - Base URL: `/api/v1`
    - Auth: `Authorization: Bearer {token}`
    - JSON: camelCase
    - Enum в JSON: lower-case строки
servers:
  - url: /api/v1
...
paths:
...
  /cars:
    get:
      tags: [Cars]
      summary: Search cars
      operationId: searchCars
      parameters:
        - in: query
          name: priceMin
          schema: { type: integer, minimum: 0, nullable: true }
        - in: query
          name: priceMax
          schema: { type: integer, minimum: 0 }
        - in: query
          name: bodyType
          schema: { $ref: '#/components/schemas/BodyType' }
        - in: query
          name: engineType
          schema: { $ref: '#/components/schemas/EngineType' }
        - in: query
          name: brand
          schema: { type: string, nullable: true }
        - in: query
          name: seats
          schema: { type: integer, nullable: true }
        - in: query
          name: transmission
          schema: { $ref: '#/components/schemas/Transmission' }
        - in: query
          name: drive
          schema: { $ref: '#/components/schemas/DriveType' }
        - in: query
          name: yearMin
          schema: { type: integer, nullable: true }
        - in: query
          name: yearMax
          schema: { type: integer, nullable: true }
        - in: query
          name: powerMin
          schema: { type: integer, minimum: 50, nullable: true }
        - in: query
          name: powerMax
          schema: { type: integer, maximum: 1000, nullable: true }
        - in: query
          name: fuelConsumptionMax
          schema: { type: number, minimum: 2, maximum: 30, nullable: true }
        - in: query
          name: engineVolumeMin
          schema: { type: number, minimum: 0.5, nullable: true }
        - in: query
          name: engineVolumeMax
          schema: { type: number, maximum: 10, nullable: true }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: perPage
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCarResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cars/{carId}:
    parameters:
      - in: path
        name: carId
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Cars]
      summary: Get car by id
      operationId: getCar
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    BodyType:
      type: string
      enum: [sedan, suv, hatchback, wagon, minivan, coupe, pickup]
    EngineType:
      type: string
      enum: [petrol, diesel, hybrid, electric]
    Transmission:
      type: string
      enum: [manual, automatic, robot, cvt]
    DriveType:
      type: string
      enum: [fwd, rwd, awd]
    Car:
      type: object
      required:
        - id
        - brand
        - model
        - year
        - price
        - bodyType
        - engineType
        - engineVolume
        - powerHp
        - transmission
        - drive
        - seats
        - fuelConsumption
        - description
        - imageUrl
      properties:
        id: { type: string, format: uuid }
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        price: { type: integer }
        bodyType: { $ref: '#/components/schemas/BodyType' }
        engineType: { $ref: '#/components/schemas/EngineType' }
        engineVolume: { type: number, nullable: true }
        powerHp: { type: integer }
        transmission: { $ref: '#/components/schemas/Transmission' }
        drive: { $ref: '#/components/schemas/DriveType' }
        seats: { type: integer }
        fuelConsumption: { type: number, nullable: true }
        description: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
    PagedCarResponse:
      type: object
      required: [items, total, page, perPage]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Car' }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        perPage: { type: integer, minimum: 1 }